+++ src/components/TimeKeeperContainer.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import TimeKeeper from './TimeKeeper'
-import { ConfigProvider, ConfigProps } from '../hooks/config-context'
-import { StateProvider } from '../hooks/state-context'
+import { ConfigProvider, ConfigProps } from '../hooks/useConfigContext'
+import { StateProvider } from '../hooks/useStateContext'
 import { TimeInput, ChangeTimeFn } from '../helpers/types'
 
 export interface Props extends ConfigProps {
