+++ src/components/TopBar.tsx
@@ -2,8 +2,8 @@ import React, { useState } from 'react'
 import { jsx } from '@emotion/core'
 
 import TimeDropdown from './TimeDropdown'
-import useTimekeeperState from '../hooks/state-context'
-import useConfig from '../hooks/config-context'
+import useTimekeeperState from '../hooks/useStateContext'
+import useConfig from '../hooks/useConfigContext'
 import * as styles from './styles/top-bar'
 import { MODE, MERIDIEM } from '../helpers/constants'
 import { isHourMode } from '../helpers/utils'
